Extract e2e data model helpers
