+++ frontend/test/__support__/e2e/cypress.js
@@ -16,6 +16,7 @@ export * from "./helpers/e2e-ad-hoc-question-helpers";
 export * from "./helpers/e2e-enterprise-helpers";
 export * from "./helpers/e2e-mock-app-settings-helpers";
 export * from "./helpers/e2e-assertion-helpers";
+export * from "./helpers/e2e-data-model-helpers";
 
 export function setupLocalHostEmail() {
   // Email info
@@ -92,17 +93,6 @@ export function setupDummySMTP() {
     "email-smtp-username": "nevermind",
     "email-smtp-password": "it-is-secret-NOT",
     "email-from-address": "nonexisting@metabase.test",
-  });
-}
-
-export function remapDisplayValueToFK({ display_value, name, fk } = {}) {
-  // Both display_value and fk are expected to be field IDs
-  // You can get them from frontend/test/__support__/e2e/cypress_sample_dataset.json
-  cy.request("POST", `/api/field/${display_value}/dimension`, {
-    field_id: display_value,
-    name,
-    human_readable_field_id: fk,
-    type: "external",
   });
 }
